Move the StorageClass away from the Helm Chart.
I created a script to create the StorageClass easily.

I updated the README to be more precise about what to do with Helm.

The StorageClass is a shared resource shared across all the OpenNMS
instances. It must exist before adding the Helm Charts.

In the future, the Helm Chart could verify if it exists before applying the
changes.