Fixing problems with external dependencies.
I refactored the Helm Chart to avoid issues when passing certificate or
JKS files content (and increased the version of the chart, to denote
this breaking change).

Updated the README to use the new method, and improved the Helm
installation/upgrade commands.

Reworked the deployment of Loki (for testing) due to major changes on
the current Helm chart.

Switched to Helm to install cert-manager (for testing).

Added simple validation to the helper scripts to ensure that all
commands used are available.