Support RPM, Debian, and tarball-based OpenNMS Horizon containers
This fixes version checking for more recent versions of OpenNMS during
the Horizon 30/31 timeframe when two subsequent changes were made:
* 30.0.0: Move to Ubuntu base image and Debian OpenNMS packages away from RPM:
  https://github.com/OpenNMS/opennms/commit/9a7a246878969cc987af6c317775e71e7a18bfbe
* 30.0.4: Move from Debian OpenNMS packages to using the tarball artifact from Maven:
  https://github.com/OpenNMS/opennms/commit/cb867870cdb2b5818a8da844943cc8314211d417

Starting in Horizon 30.0.4, the container build process changed and
installs tarballs instead of packages, so we can no longer get the
package type (Horizon/Meridian) and version from the package manager.
Get the version from jetty-webapps/opennms/WEB-INF/version.properties
in this case and mark the package name as "unknown" for now.

Alternatively, we could drop the check for Kafka Twin support and
assume it is always there now in modern versions of OpenNMS, but I
figure we'll need to some some version checks in the future, so I
kept it for now.

Good details from Alejandro on the history behind the version checks
and possible approaches are in this comment:
https://github.com/opennms-forge/onms-k8s-poc/pull/6#issuecomment-1276253630