Support admin password changes and restart onms-core on upgrades
Previously, the OpenNMS HTTP admin password was purposefully only
set on initial install. This has changed and now the admin password
is updated to what is set in the Helm chart every time the onms-core
pod is restarted (in particular: when the init container for the pod
runs).

For the above work, a Perl script was added that edits users.xml and
it makes a decent attempt to properly edit users.xml, but it might
run into problems if the user whose password is being changed doesn't
have a password element in users.xml. This functionality really should
be replaced with a proper Java implementation in OpenNMS. Also, the
Perl script only does MD5 hashing on the admin password, compared to
the much stronger SHA-256 hashing with many rounds normally done,
which is another reason to add CLI functionality to OpenNMS to do
this properly with OpenNMS' UserFactory.

Also, to ensure the password will always be updated, we need to make
sure the init container always runs when the Helm chart is upgraded.
To do this, a "rollme" metadata annotation was added, following the
pattern in the Helm documentation. This has the nice affect of
ensuring that many other changes are applied, too, without requiring
manual work.
See: https://v3.helm.sh/docs/howto/charts_tips_and_tricks/#automatically-roll-deployments